requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.124.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose
bcrypt==3.2.2
python-multipart